requirements.txt
gradio
google-generativeai
nest_asyncio
